backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiofiles==23.2.1

# Image/Video Processing
pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.26.2

# ML + Deepfake Detection Dependencies
scikit-learn==1.3.2

# PyTorch (CPU wheels; good starting point)
torch==2.1.1
torchvision==0.16.1

# Face recognition utilities
facenet-pytorch==2.5.3

# TensorFlow (works only on Python 3.10)
tensorflow==2.15.0

# Auth / JWT
python-jose==3.3.0

passlib[bcrypt]==1.7.4

